+++ db/drone_order_dao.py
@@ -1,5 +1,5 @@
 from db.database import fetchall, change
-from db.dos import DroneOrder, map_to_objects
+from db.data_objects import DroneOrder, map_to_objects
 
 from typing import List
